backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via edpop-explorer
asgiref==3.8.1
    # via django
beautifulsoup4==4.13.3
    # via -r requirements.in
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
cmd2==2.5.11
    # via edpop-explorer
defusedxml==0.7.1
    # via sruthi
dj-rest-auth==7.0.1
    # via -r requirements.in
django==4.2.19
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-filter
    #   djangorestframework
    #   restframework-rdf
django-filter==25.1
    # via -r requirements.in
django-json-widget==2.0.1
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   restframework-rdf
edpop-explorer==0.14.1
    # via -r requirements.in
flatten-dict==0.4.2
    # via sruthi
idna==3.10
    # via requests
iso639-lang==2.6.0
    # via edpop-explorer
isodate==0.7.2
    # via rdflib
lxml==5.3.1
    # via -r requirements.in
psycopg==3.2.5
    # via -r requirements.in
pygments==2.19.1
    # via edpop-explorer
pyparsing==3.2.1
    # via rdflib
pyperclip==1.9.0
    # via cmd2
pyyaml==6.0.2
    # via edpop-explorer
rdflib==7.1.3
    # via
    #   -r requirements.in
    #   edpop-explorer
    #   restframework-rdf
    #   sparqlwrapper
requests==2.32.3
    # via
    #   -r requirements.in
    #   sruthi
restframework-rdf==1.2.0
    # via -r requirements.in
six==1.17.0
    # via flatten-dict
soupsieve==2.6
    # via beautifulsoup4
sparqlwrapper==2.0.0
    # via edpop-explorer
sqlparse==0.5.3
    # via django
sruthi==2.0.0
    # via
    #   -r requirements.in
    #   edpop-explorer
termcolor==2.5.0
    # via edpop-explorer
typing-extensions==4.12.2
    # via
    #   asgiref
    #   beautifulsoup4
    #   edpop-explorer
    #   psycopg
urllib3==2.3.0
    # via requests
wcwidth==0.2.13
    # via cmd2
xmltodict==0.14.2
    # via
    #   edpop-explorer
    #   sruthi
